requirements.txt
Django==2.1.5
django-filter==2.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
Markdown==3.0.1
Pillow==5.4.1
pytz==2018.9
